Вывод времени последнего обновления

+++ app/src/main/java/com/alexd/weather/model/Location.java
@@ -7,23 +7,23 @@ import java.io.Serializable;
  */
 public class Location implements Serializable {
 
-    private float longitude;
-    private float latitude;
+    private double longitude;
+    private double latitude;
     private long sunset;
     private long sunrise;
     private String country;
     private String city;
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
     public long getSunset() {
